+++ RobBiped/RobBiped/Utils/ExponentialFilterWithPeakRejection.cpp
@@ -85,7 +85,7 @@ double ExpFilterPeakReject::filter_pr(int16_t raw_value, bool accept_step)
 
 double ExpFilterPeakReject::filter_pr(uint32_t raw_value, bool accept_step)
 {
-	if (abs(abs(raw_value) - abs(last_filtered_value_) > threshold_value_) && !(accept_step))
+	if (abs(abs((int32_t)raw_value) - abs((int32_t)last_filtered_value_) > threshold_value_) && !(accept_step))
 	{
 		return last_filtered_value_;
 	}
@@ -109,7 +109,7 @@ double ExpFilterPeakReject::filter_pr(int32_t raw_value, bool accept_step)
 
 double ExpFilterPeakReject::filter_pr(uint64_t raw_value, bool accept_step)
 {
-	if (abs(abs(raw_value) - abs(last_filtered_value_) > threshold_value_) && !(accept_step))
+	if (abs(abs((int32_t)raw_value) - abs((int32_t)last_filtered_value_) > threshold_value_) && !(accept_step))
 	{
 		return last_filtered_value_;
 	}
